refactor(stores): split hud.svelte.ts into focused sub-stores
